Client-side OAuth part 2
The backend service has a grpc interceptor for fetching auth metadata,
validating the token, and appending to the context for handlers to use.

The frontend handles posting to the OAuth API based on configuration
loaded from the backend. The token received is then validated and stored
in localStorage.

A ReplaySubject on a shared service would be useful here to stop our
app from re-doing requests all over the place.

It would also be nice if the front end knew when the token expired, and
did a proper logout or refresh.

Issue #1